+++ api/src/main/kotlin/app/meetacy/backend/mock/storage/MockToken.kt
@@ -1,6 +1,6 @@
 package app.meetacy.backend.mock.storage
 
 data class MockToken(
-    val ownerId: Int,
+    val ownerId: Long,
     val value: String
 )
